Use stemming in post_search
Stemming is the process of reducing words
to their word stem, base, or root form.
Stemming is used by search engines to reduce indexed words
to their stem, and to be able to match inflected or derived words.
For example, "music" and "musician"
can be considered similar words by a search engine.